Added StopDataFetcher for grabbing stop data
StopDataFetcher can get information about a specific stop and return it in a StoInfo struct, or search for stops and return a list of StopInfo in a range of GPS coordinates. Used Linq's xml parsing capabilities and also changed the departure data fetching in DataParser to use Linq. Changed Departure to a struct because it seemed more fitting.